commands: move catalog subcommands to subdirs
