+++ command/catalog/catalog_test.go
@@ -1,4 +1,4 @@
-package cat
+package catalog
 
 import (
 	"strings"
