+++ command/catalog/list/dc/catalog_list_datacenters.go
@@ -1,4 +1,4 @@
-package catlistdc
+package dc
 
 import (
 	"flag"
