+++ command/catalog/list/dc/catalog_list_datacenters_test.go
@@ -1,4 +1,4 @@
-package catlistdc
+package dc
 
 import (
 	"strings"
