+++ command/catalog/list/nodes/catalog_list_nodes.go
@@ -1,4 +1,4 @@
-package catlistnodes
+package nodes
 
 import (
 	"flag"
