+++ command/catalog/list/services/catalog_list_services.go
@@ -1,4 +1,4 @@
-package catlistsvc
+package services
 
 import (
 	"bytes"
